Move functions and change math to numpy
