+++ pyart/retrieve/__init__.py
@@ -11,7 +11,6 @@ from .echo_class import hydroclass_semisupervised  # noqa
 from .echo_class import steiner_conv_strat  # noqa
 from .gate_id import fetch_radar_time_profile, map_profile_to_gates  # noqa
 from .kdp_proc import kdp_maesaka, kdp_schneebeli, kdp_vulpiani  # noqa
-from .precip_rate import ZtoR  # noqa
 from .qpe import est_rain_rate_a  # noqa
 from .qpe import est_rain_rate_hydro  # noqa
 from .qpe import est_rain_rate_kdp  # noqa
@@ -19,6 +18,7 @@ from .qpe import est_rain_rate_z  # noqa
 from .qpe import est_rain_rate_za  # noqa
 from .qpe import est_rain_rate_zkdp  # noqa
 from .qpe import est_rain_rate_zpoly  # noqa
+from .qpe import ZtoR  # noqa
 from .qvp import quasi_vertical_profile  # noqa
 from .simple_moment_calculations import calculate_snr_from_reflectivity  # noqa
 from .simple_moment_calculations import calculate_velocity_texture  # noqa
