+++ tests/retrieve/test_qpe.py
@@ -1,6 +1,7 @@
 """ Unit tests for rainfall rate estimation module. """
 
 from numpy.testing import assert_allclose
+import numpy as np
 
 import pyart
 
@@ -153,3 +154,27 @@ def test_get_coeff_rkdp():
 
     coeff_rkdp_use_x = pyart.retrieve.qpe._get_coeff_rkdp(13e9)
     assert coeff_rkdp_use_x == (15.81, 0.7992)
+
+def test_precip_rate():
+    grid = pyart.testing.make_storm_grid()
+    grid = pyart.retrieve.ZtoR(grid)
+
+    # check that field is in grid object
+    assert "NWS_primary_prate" in grid.fields.keys()
+
+    # check calculations are within 10^-4 orders of magnitude
+    assert (
+        np.floor(
+            np.log10(
+                grid.fields["NWS_primary_prate"]["data"][0, 10, 10]
+                - (
+                    (
+                        (10 ** (grid.fields["reflectivity"]["data"][0, 10, 10] / 10))
+                        / 300
+                    )
+                    ** (1 / 1.4)
+                )
+            )
+        )
+        < -4
+    )
